pyproject.toml
[project]
name = "degradation-htr-test"
version = "0.1.0"
description = "Minimal test for degradation-aware HTR"
requires-python = ">=3.10"
dependencies = [
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "pillow>=10.0.0",
    "jiwer>=3.0.0",
    "numpy>=2.2.6",
    "matplotlib>=3.10.7",
    "networkx>=3.4.2",
    "scipy>=1.15.3",
    "scikit-image>=0.25.2",
    "ipython>=8.37.0",
    "imageio>=2.37.2",
    "pandas>=2.3.3",
    "pykan>=0.2.8",
    "scikit-learn>=1.7.2",
    "tensorflow>=2.20.0",
    "opencv-contrib-python>=4.11.0.86",
    "pytorch-lightning>=2.5.6",
    "lightning>=2.5.6",
    "timm>=1.0.22",
    "albumentations>=2.0.8",
    "nibabel>=5.3.2",
    "h5py>=3.15.1",
    "pygrappa>=0.26.3",
    "phantominator>=0.7.0",
    "sigpy>=0.1.27",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "matplotlib>=3.7.0",
]
